Cygwin needs init_Win32CORE called early.
This needs cleaning up. The configuration logic needs to be tidier (atm
we lose NEED_PERLAPI_H on Cygwin, which isn't important but should be
fixed). I need to check that calling init_Win32CORE like this is
actually the right thing to do: how does core perl manage to get an
init_Win32CORE symbol in the main exe?